# Use strings.Builder in lexer