@pelletier 2020-09-12
That's awesome! Thanks so much for doing that!

As far as memory consumption goes: I think it's a fundamental issue in the way `(Un)Marshal` was written.

For unmarshalling, it always builds a `Tree` first. I think the parser should be changed to accept a "builder" interface, with two implementation: one that builds a `Tree`, and one that unmarshals directly. Tried to do that [at some point](https://github.com/pelletier/go-toml/tree/unmarshal-rewrite), but not enough time to dedicate to that rewrite at the moment.

For marshalling, I think building the Tree should also be bypassed. That one is probably less effort because it's mostly a matter of isolating the conversion functions.